refactor: 커스텀훅 리팩토링 (#64)
